Bump to 0.2.0
