+++ .gitignore
@@ -12,3 +12,4 @@
 
 vendor/bundle
 Gemfile.lock
+.ruby-version
