+++ lib/open_api_2_json_schema/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module OpenApi2JsonSchema
-  VERSION = "0.1.2"
+  VERSION = "0.2.0"
 end
